+++ lib/model/hazard_type.dart
