+++ lib/model/hazard_update.dart
