+++ lib/model/snapped_latlng.dart
