+++ lib/model/trail.dart
