+++ lib/page/common/add_hazard_modal.dart
@@ -4,14 +4,14 @@ import 'dart:io';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
-import 'package:forest_park_reports/models/hazard.dart';
-import 'package:forest_park_reports/models/hazard_type.dart';
-import 'package:forest_park_reports/pages/home_page/panel_page.dart';
-import 'package:forest_park_reports/providers/hazard_provider.dart';
-import 'package:forest_park_reports/providers/location_provider.dart';
-import 'package:forest_park_reports/providers/trail_provider.dart';
+import 'package:forest_park_reports/model/hazard.dart';
+import 'package:forest_park_reports/model/hazard_type.dart';
+import 'package:forest_park_reports/page/home_page/panel_page.dart';
+import 'package:forest_park_reports/provider/hazard_provider.dart';
+import 'package:forest_park_reports/provider/location_provider.dart';
+import 'package:forest_park_reports/provider/trail_provider.dart';
 import 'package:forest_park_reports/util/extensions.dart';
-import 'package:forest_park_reports/util/permissions_dialog.dart';
+import 'package:forest_park_reports/page/common/permissions_dialog.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:image_picker/image_picker.dart';
