+++ lib/page/common/statusbar_blur.dart
@@ -1,7 +1,6 @@
 import 'dart:ui';
 
 import 'package:flutter/widgets.dart';
-import 'package:forest_park_reports/consts.dart';
 
 class StatusBarBlur extends StatelessWidget {
   const StatusBarBlur({super.key});
