+++ lib/provider/map_cursor_provider.dart
