+++ lib/util/fl_latlng_spot.dart
@@ -1,7 +1,7 @@
 import 'package:fl_chart/fl_chart.dart';
 import 'package:latlong2/latlong.dart';
 
-import '../models/hazard.dart';
+import '../model/hazard.dart';
 
 class FlCoordinateSpot extends FlSpot {
   final LatLng position;
